// src/interfaces/boleto.interface.ts
export interface IntBoleto {
  informativo: string;
  logoBanco: string;
  codigoDoBanco: string;
  localDePagamento: string;
  beneficiario: string;
  dataDoDocumento: string;
  numeroDocumento: string;
  aceite: string;
  dataProcessamento: string;
  usoDoBanco: string;
  carteira: string;
  especieDocumento: string;
  especie: string;
  quantidade: string;
  valor: string;
  instrucaoDaCompensacao: string;
  instrucaoDoSacado: string;
  pagador: string;
  sacadorAvalista: string;
  codigoBaixa: string;
  enderecoSacado: string;
  vencimento: string;
  agenciaCodigoBeneficiario: string;
  nossoNumero: string;
  valorDocumento: number;
  descontoAbatimento: string;
  outrasDeducoes: string;
  jurosMulta: string;
  outrosAcrescimos: string;
  valorCobrado: string;
  codigoDeBarras: string;
  linhaDigitavel: string;
  nomeSacadoCompleto: string
}

export interface BoletoCarne {
  informativo: string;
  logoBanco: string;
  codigoDoBanco: string;
  localDePagamento: string;
  beneficiario: string;
  dataDoDocumento: string;
  numeroDocumento: string;
  aceite: string;
  dataProcessamento: string;
  usoDoBanco: string;
  carteira: string;
  especieDocumento: string;
  especie: string;
  quantidade: string;
  valor: string;
  instrucaoDaCompensacao: string;
  instrucaoDoSacado: string;
  pagador: string;
  sacadorAvalista: string;
  codigoBaixa: string;
  enderecoSacado: string;
  vencimento: string;
  agenciaCodigoBeneficiario: string;
  nossoNumero: string;
  valorDocumento: number;
  descontoAbatimento: string;
  outrasDeducoes: string;
  jurosMulta: string;
  outrosAcrescimos: string;
  valorCobrado: string;
  codigoDeBarras: string;
  linhaDigitavel: string;
  nomeSacadoCompleto: string;
  enderecoSacadoCarne: string;
  bairroSacado: null;
  enderecoSacadoCont: string;
}